server-requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
python-multipart>=0.0.9
websockets>=12.0
bcrypt>=4.0.0
